Nem sei o que mudei

+++ 4/rootfs/large.c
@@ -16,7 +16,7 @@ int main()
 		vector[i] = 12345;
 	}
 
-	printf("ponteiro para o vetor: %lld\n", (long long)vector);
+	printf("ponteiro para o vetor: %ld\n", (unsigned long)vector);
 	while(1);
 
 	return 0;
